Fix double-dependencies
If you add package A which depends on Z, and package B which depends on
Z, then during the build process it would try to build the package Z
twice but it would complain because the directory already exists and was
built as part of this build already.

This patch passes the 'force' parameter to cower to alleviate the issue.